requirements.txt
yfinance>=0.2.28
supabase>=2.3.4
python-dotenv>=1.0.0
pandas>=2.2.0
langchain-community>=0.0.20
langchain-groq>=0.0.1
crewai>=0.28.0
groq>=0.4.0 